layout components for page header and footer
