+++ app/components/layout/index.ts
@@ -1,0 +1,2 @@
+export { PageFooter } from './page-footer'
+export { PageHeader, PageHeaderDev } from './page-header'
